style(commands): copy buttons
