+++ components/CopyButton.jsx
@@ -1,6 +1,8 @@
 import { useState } from "react";
-import { IconButton } from "./Typography";
+import { IconButton } from "./Buttons";
 import { CopyIcon, CheckIcon } from "@radix-ui/react-icons";
+import { HStack } from "@chakra-ui/react";
+import { Text } from "./Typography";
 
 const CopyButton = ({ value }) => {
     const [copied, setCopied] = useState(false);
@@ -12,8 +14,11 @@ const CopyButton = ({ value }) => {
     };
 
     return (
-        <IconButton onClick={handleCopy}>
-            {copied ? <CheckIcon /> : <CopyIcon />}
+        <IconButton onClick={handleCopy} aria-label={copied ? "Copied" : "Copy to clipboard"}>
+            <HStack spacing={1}>
+                {copied ? <CheckIcon width={20} height={20} /> : <CopyIcon width={20} height={20} />}
+                <Text>{copied ? "Copied!" : "Copy"}</Text>
+            </HStack>
         </IconButton>
     );
 };
